+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 Changelog
 =========
 
-0.0.1 (2019-11-21)
+0.0.1 (2019-11-22)
 -------------------
 
+Added activation(), downScale(), conformity(), pooling(), normalize() functions
+Create main class
 Initial commit